@@ -5,7 +5,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>3.2.4</version>
+		<version>3.4.5</version>
 		<relativePath/> <!-- lookup children from repository -->
 	</parent>
 	<groupId>edu.harvard.dbmi.avillach</groupId>
@@ -14,7 +14,7 @@
 	<name>dump</name>
 	<description>Data Dictionary Dump API</description>
 	<properties>
-		<java.version>21</java.version>
+		<java.version>24</java.version>
 		<spotless.version>2.41.1</spotless.version>
 	</properties>
 	<dependencies>
@@ -44,7 +44,6 @@
 		<dependency>
 			<groupId>org.springframework.data</groupId>
 			<artifactId>spring-data-commons</artifactId>
-			<version>3.3.0</version>
 		</dependency>
 
 		<dependency>
