@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.http.ResponseEntity;
 
 import java.time.LocalDateTime;
@@ -18,7 +18,7 @@ class DumpControllerTest {
     @Autowired
     DumpController subject;
 
-    @MockBean
+    @MockitoBean
     DumpService service;
 
     @Test
