@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
 import java.time.LocalDateTime;
 
@@ -15,7 +15,7 @@ class DumpServiceTest {
     @Autowired
     DumpService subject;
 
-    @MockBean
+    @MockitoBean
     DumpRepository repository;
 
     @Test
