@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
 import java.util.List;
 import java.util.Optional;
@@ -13,10 +13,10 @@ import java.util.Optional;
 @SpringBootTest
 class DataRefreshServiceTest {
 
-    @MockBean
+    @MockitoBean
     RemoteDictionaryRepository repository;
 
-    @MockBean
+    @MockitoBean
     RemoteDictionaryAPI api;
 
     @Autowired
