@@ -12,7 +12,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
@@ -27,8 +27,10 @@ import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.*;
 class RemoteDictionaryAPITest {
 
     private final DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS");
-    private final ObjectMapper mapper = new ObjectMapper().setVisibility(PropertyAccessor.FIELD, ANY);
-    @MockBean
+    @Autowired
+    ObjectMapper mapper;
+
+    @MockitoBean
     CloseableHttpClient client;
 
     @Autowired
