@@ -1,6 +1,6 @@
 package edu.harvard.dbmi.avillach.dump.remote;
 
-import edu.harvard.dbmi.avillach.dump.entities.ConceptNodeDump;
+import edu.harvard.dbmi.avillach.dump.entities.*;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -71,11 +71,75 @@ class RemoteDictionaryRepositoryInsertTest {
         Assertions.assertEquals(7, gicConcepts);
     }
 
+    @Test
+    void shouldAddFacets() {
+        FacetDump child = new FacetDump("foo", "Foo", "fooer", "study_ids_dataset_ids", 1, null);
+        FacetDump parent = new FacetDump("fooz", "Fooz", "fooerz", "study_ids_dataset_ids", List.of(child), 1, null);
+
+        Integer before = template.queryForObject("SELECT count(*) FROM facet", Integer.class);
+        subject.addFacetsForSite("bch", List.of(parent, parent));
+        Integer after = template.queryForObject("SELECT count(*) FROM facet", Integer.class);
+        Assertions.assertEquals(before + 2, after);
+    }
+
+    @Test
+    void shouldAddFacetCategory() {
+        FacetCategoryDump cat = new FacetCategoryDump("name", "display", "description");
+
+        Integer before = template.queryForObject("SELECT count(*) FROM facet_category", Integer.class);
+        subject.addFacetCategoriesForSite("bch", List.of(cat, cat));
+        Integer after = template.queryForObject("SELECT count(*) FROM facet_category", Integer.class);
+        Assertions.assertEquals(before + 1, after);
+    }
+
+    @Test
+    void shouldAddFacetCategoryMeta() {
+        FacetCategoryMetaDump meta = new FacetCategoryMetaDump("study_ids_dataset_ids", "key", "value");
+
+        Integer before = template.queryForObject("SELECT count(*) FROM facet_category_meta", Integer.class);
+        subject.addFacetCategoryMetasForSite("bch", List.of(meta, meta));
+        Integer after = template.queryForObject("SELECT count(*) FROM facet_category_meta", Integer.class);
+        Assertions.assertEquals(before + 1, after);
+    }
+
+    @Test
+    void shouldAddFacetMeta() {
+        FacetMetaDump meta = new FacetMetaDump("phs002715", "study_ids_dataset_ids", "key", "value");
+
+        Integer before = template.queryForObject("SELECT count(*) FROM facet_meta", Integer.class);
+        subject.addFacetMetasForSite("bch", List.of(meta, meta));
+        Integer after = template.queryForObject("SELECT count(*) FROM facet_meta", Integer.class);
+        Assertions.assertEquals(before + 1, after);
+    }
+
+    @Test
+    void shouldPair() {
+        FacetConceptPair pair = new FacetConceptPair("phs002715", "study_ids_dataset_ids", "\\Consent Type\\GIC Consent\\");
+
+        Integer before = template.queryForObject("SELECT count(*) FROM facet__concept_node", Integer.class);
+        subject.addFacetConceptPairsForSite("bch", List.of(pair, pair));
+        Integer after = template.queryForObject("SELECT count(*) FROM facet__concept_node", Integer.class);
+        Assertions.assertEquals(before + 1, after);
+    }
+
+    @Test
+    void shouldAddConceptMetas() {
+        fixWeirdSeedBug();
+        ConceptNodeMetaDump pair = new ConceptNodeMetaDump("\\Consent Type\\GIC Consent\\", "key123123", "value");
+
+        Integer before = template.queryForObject("SELECT count(*) FROM concept_node_meta", Integer.class);
+        subject.addConceptMetasForSite("bch", List.of(pair, pair));
+        Integer after = template.queryForObject("SELECT count(*) FROM concept_node_meta", Integer.class);
+        Assertions.assertEquals(before + 1, after);
+    }
+
     private void fixWeirdSeedBug() {
         List<String> queries = List.of(
             "INSERT INTO dataset (REF, FULL_NAME, ABBREVIATION) VALUES ('test', 'test', 'test')",
             "INSERT INTO concept_node (DATASET_ID, NAME, DISPLAY, CONCEPT_TYPE, CONCEPT_PATH) VALUES (23, 'test', 'test', 'continuous', '\\a')",
-            "INSERT INTO concept_node (DATASET_ID, NAME, DISPLAY, CONCEPT_TYPE, CONCEPT_PATH) VALUES (23, 'test', 'test', 'continuous', '\\a')"
+            "INSERT INTO concept_node (DATASET_ID, NAME, DISPLAY, CONCEPT_TYPE, CONCEPT_PATH) VALUES (23, 'test', 'test', 'continuous', '\\a')",
+            "INSERT INTO concept_node_meta (CONCEPT_NODE_ID, KEY, VALUE) VALUES (216, 'K', 'V')",
+            "INSERT INTO concept_node_meta (CONCEPT_NODE_ID, KEY, VALUE) VALUES (216, 'K', 'V')"
         );
 
         for (String query : queries) {
